+++ src/inkling/frame.rs
@@ -13,3 +13,31 @@ use smash::app::*;
 use smash::phx::Vector3f;
 use crate::util::*;
 use super::*;
+pub fn install() {
+    smashline::install_agent_frame_callbacks!(ink);
+}
+#[fighter_frame_callback]
+pub fn ink(fighter : &mut L2CFighterCommon) {
+    unsafe {
+        let boma = smash::app::sv_system::battle_object_module_accessor(fighter.lua_state_agent); 
+		let status_kind = smash::app::lua_bind::StatusModule::status_kind(boma);
+		let fighter_kind = smash::app::utility::get_kind(boma);
+		let lua_state = fighter.lua_state_agent;
+		let ENTRY_ID = WorkModule::get_int(boma, *FIGHTER_INSTANCE_WORK_ID_INT_ENTRY_ID) as usize;
+		let motion_kind = MotionModule::motion_kind(boma);
+		let frame = MotionModule::frame(boma);
+		if fighter_kind == *FIGHTER_KIND_INKLING {
+			let cat1 = ControlModule::get_command_flag_cat(boma, 0);
+			if status_kind == *FIGHTER_INKLING_STATUS_KIND_CHARGE_INK {
+				if KineticModule::get_kinetic_type(boma) != *FIGHTER_KINETIC_TYPE_DASH && (cat1 & *FIGHTER_PAD_CMD_CAT1_FLAG_DASH) != 0{
+					StatusModule::change_status_request_from_script(boma, *FIGHTER_INKLING_STATUS_KIND_CHARGE_INK_END, true);
+					KineticModule::change_kinetic(boma, *FIGHTER_KINETIC_TYPE_DASH);
+				};
+				if KineticModule::get_kinetic_type(boma) != *FIGHTER_KINETIC_TYPE_DASH && (cat1 & *FIGHTER_PAD_CMD_CAT1_FLAG_TURN_DASH) != 0{
+					StatusModule::change_status_request_from_script(boma, *FIGHTER_INKLING_STATUS_KIND_CHARGE_INK_END, true);
+					KineticModule::change_kinetic(boma, *FIGHTER_KINETIC_TYPE_DASH_BACK);
+				};
+			};
+		};
+	};
+}
